Also check on start
