+++ packages/node/src/behaviors/operational-credentials/index.ts
@@ -9,4 +9,3 @@
 export * from "./OperationalCredentialsInterface.js";
 export * from "./OperationalCredentialsBehavior.js";
 export * from "./OperationalCredentialsServer.js";
-export * from "./FabricScopedDataHandler.js";
